prefix pex/par running of python command with arch -x86_64
Summary:
This is a stop gap until we figure out how to plumb the build arch for macos down to these levels.   But at that point we will pass x86_64 or arm64

But this forces the mac to execute python3.10 in x86_64 emulation since the build artifacts for now are by default that platform.

Reviewed By: itamaro

Differential Revision: D43472681

fbshipit-source-id: fc5a902c5a27e85ab59f8c52edff0312572a95de