Add support/hooks for user-provided cycle detection
Summary:
For the short term, it's critical that we detect cycles in loads and configured
target graph. To do that, I'll add some targeted cycle detection.

Originally I wanted to do this cycle detection outside of dice. This hit (at
least) three problems:

1. the cycle detection necessarily was informed of edges before dice was
informed of them. if a cycle was detected it could return an error without
dice actually learning about the dep that caused the error. if the error
was not marked "invalid" by the key type, dice could cache the error and not
invalidate it when the bad dep was fixed.
2. it requires explicitly marking around every key and dependency edge. for
load this is fine because it's just one spot. for configured graph it technically
should track for cycles through like some transition keys and configuration
node keys and toolchain keys etc.
3. on node recompute, dice will wait on its known dep edges with no way for
something external to be aware of it.

(1) is generally resolvable.
(2) was a pain, but just marking the main configured graph edges would get the
vast majority of cycles, especially those for end users (rather than
integrators, rule authors, etc).
(3) was a total blocker

For resolving (3) there's a couple options:

1. allow Keys to say they cannot be recomputed
2. add dice hook to tell something else when it is recomputing a key
and what the deps are
3. add dice hooks to inform something about all computation start/end and edges
as dice does for its own cycledetection (approximately, its own doesn't
actually do start/end)

Of those, (1) is easiest, but is kinda the worst of those solutions. (2) is
also like a not great solution and requires a lot of the work needed for (3)
(for example, the stuff for converting keys into `&dyn Any` that user code can
 then interpret). (3) leads to a solution that can solve both (2)+(3) problems
above.

A final alternative would be to just do cycle detection fully in dice. The main
problem with that is that doing full cycle detection in dice has very stringent
perf/mem/etc requirements.

So, this change adds the hooks needed to support custom user-provided cycle
detection.

Reviewed By: krallin

Differential Revision: D43342213

fbshipit-source-id: 06e544d4e7414470056b7e4c35bf45c63503f7dc