Add lazy CycleDetector
Summary:
This adds a "lazy" cycle detector. This will track node computations
starting+dependency+finished. After an idle period it will the traverse the
active graph and identify any cycles. Nodes in a cycle will be marked as in
the cycle. If they use the future-wrapper `guard_this` while waiting on
a dependency edge, it will exit with a cycle error.

This just provides sort of the easiest to integrate cycle detection to handle
the critical cases of load and configured graph cycles.

Initially I implemented this without the mapping of keys to u32 (i.e. just
hold keys everywhere), but that had a more significant memory cost.

Reviewed By: krallin

Differential Revision: D43342212

fbshipit-source-id: 123896e483568723f442e818eb10f152de6f1e94