Parallelize calls to get_template_info_provider_artifacts
Summary: Currently, we invoke `get_template_info_provider_artifacts` asynchronously in a loop. This is OK because at this point in code, we've already done the analysis for all of our deps, so `get_template_info_provider_artifacts` doesn't have much async work to actually do. However, for `audit classpath`, we are going to invoke this directly without having to invoke `analysis` of our deps, so let's make this concurrent.

Reviewed By: krallin

Differential Revision: D43562680

fbshipit-source-id: 7fe0b7a61efc8763b6b916f5d83cb22ace8babb8