buck2: materializer: don't use `Instant` for action ages
Summary:
The comment in the code explains it all. On MacOS, `Instant:elapsed()` appears
to not increment across sleeps, so we end up thinking a few minutes have
elapsed when in reality a few hours have.

Using wall time has its own issues with leap seconds or clock drift, but those
are a) rare events and b) are unlikely to cause *substantial* changes in the
clock, whereas putting your laptop to sleep happens a) every day and b) is
often a multi-hour affair.

The reason we don't care about small clock drift is allow for some leeway in
determining if things are expired in places where it matters and just re-check
them again (albeit not in the specific code this diff changes, because that is
only used for logging, and in fact the "runtime" logic doesn't have this issue
because it already used `DateTime<Utc>` and not `Instant`).

Reviewed By: ndmitchell

Differential Revision: D43563383

fbshipit-source-id: edf76e01f95bce6acdd7652421ddb75d6e1e89f4