buck2: convert targets to write using PartialResult
Summary:
NOTE: See D43544865 for rationale

Like it says in the title. The end goal of this stack is to completely
remove `RawOutput` (and `LspMessage`) from buck2_data, and for all command
output to go through this.

Note that this also allows us to return structured data to the
client instead of just stdout bytes, but this diff does not do that.

Reviewed By: bobyangyf

Differential Revision: D43568721

fbshipit-source-id: 951e3a505bc0604368b02cf96c1c4778afa94099