buck2: migrate Lsp to PartialResult
Summary:
NOTE: See D43544865 for rationale

Here as well, we were sending structured data via our event stream and
pretending it was just an event and handling it in the console, which is a
clever hack but a bit of an awkward one (for reasons detailed earlier on this
stack).

Now that we have a framework to send this as a partial response (i.e. a legit
streaming API that isn't actually streaming events + a response)), let's just
do so.

Reviewed By: stepancheg

Differential Revision: D43569474

fbshipit-source-id: 92edd325439c4e545550bf06c79e940b3146a658