buck2: update client to expect partial results
Summary:
NOTE: See D43544865 for rationale

Like it says in the title. This stops one step short of actually having the
PartialResultHandler finalize with the actual CommandResult, which is probably
what ultimately this should be, but for now since our only partial results
are gonna be stdout output, this will be fine.

Reviewed By: stepancheg

Differential Revision: D43544862

fbshipit-source-id: 1c49da9947d8f355c02646fe7b0c5ebab20fa9d9