Add event sink statistics to invocation recorder
Summary:
This will add the statistics of success/failure/dropped and max queue length of daemon's event sink to Invocation Record. It makes use of the counters in SnapShot, so we should note that this method is not perfect; it isn't exactly correct when e.g. some events go ahead of the first SnapShot, multiple commands run concurrently etc. The main purpose is to catch the tendency of how large amount the sink client dropped/queued so that we can optimize the client's parameter such as buffer size and batch size.

// Possibly we could store a map {uuid -> counters} in `producer.rs` to distinguish the counters between each command invocation, and wait for all events getting processed so that we can know which result (success/drop/fail) the events in the queue is going into. But it's not worthy for client to wait for what we don't need that perfection on.

Differential Revision: D43613936

fbshipit-source-id: 0c8c92341225460840dfdb80c8467a930da3df95