buck2: events: allow creating spans without going directly to `call_in_span`
Summary:
Like it says in the title. I need this later in this stack. We should mostly
minimize use of this (because the CPU time calculations are going to be off if
we cannot wrap execution), but sometimes it can be necessary when you don't
know exactly when your span should start.

Note that the event is still entered when it's created. This matters because we
guarantee parent spans have been emitted when we emit children spans.

Differential Revision: D43702185

fbshipit-source-id: ad9cd256d4a8ac34712f3cad348eaa6417e62bd3