Implement the entire dice task behaviour
Summary:
Implement most of the DiceTask

The implementation is approximately a copy of `Shared` and `Weak` from std, but with some custom special record keeping.

`std::future::Weak` is akin to `DiceTask`, and each `DicePromise` is a strong reference to it akin to a `std::future::Shared`.

Namely:
1. The DiceTask is always completed by a thread whose future is the JoinHandle. The thread reports updates to the state of the future via DiceTaskHandle. Simplifying the future implementation in that no `poll` will ever be doing real work. No Wakers sleeping will be awoken unless the task is ready.

2. The task is not the "standard states" of Pending, Polling, etc as stored by Shared future, but instead we store the Dice specific states so that its useful when we dump the state.

3. Wakers are tracked with their corresponding DiceKeys, allowing us to track the rdeps and see which key is waiting on what

4. We can explicitly track cancellations by tracking the Waker drops.

## Size differences:

DiceTask <-> Weak: DiceTask holds an extra `JoinHandle` which is a single ptr.  DiceTask now holds a triomphe::Arc instead of std::Arc which is slightly more efficient as it doesn't require weak ptr handling. This is just so that we have the JoinHandle so we can `abort` when canceled, but we could choose to change the implementation by moving cancellation notification into the DiceTaskInternal

DicePromise <-> Shared: DicePromise is 2 triomphe::Arc and a usize, whereas `Shared` is a usize and 1 std::Arc, so we hold one extra Arc.

Waker tracking:
We now hold a pair DiceKey and Waker, which requires DiceKey per waker. The Waker itself is now an AtomicWaker, which is an extra AtomicUsize.

So, we marginally increase the overall size of the future by 1 pointer to do explicit Key tracking.
This sounds okay.
Especially since performance wise, we reduce one layer of Arc, get rid of std in favour of triomphe Arc, and reduce wake ups of dependant keys by making the wakes of the computing thread separate from waiting threads.

Reviewed By: cjhopman

Differential Revision: D43488520

fbshipit-source-id: e91fc3ad1c3bc7002e8ef9cbc9a2c830296e2fb2