network stats: Count # bytes and digests uploaded
Summary:
This tracks the number of uploaded bytes and number of uploaded digests from the RE upload stage..

This is done as requested in T146396459 so we can understand how much data we're uploading during the course of fully-cached builds in iOS on laptops.

Note this is sort of explicitly moving away from the ExecutorStage reporting; we now no longer have a nested ReStage for this phase, but it winds up working a bit better since we can more easily track these spans and won't have to plumb through changes to `ExecutorStageEnd` everywhere we use it.

Reviewed By: krallin

Differential Revision: D43726415

fbshipit-source-id: 47b594feb10c1b94d51469dd4f12f678df5e452a