Create modern copy of graph storage
Summary:
Same as existing dice except:
1. we remove transient values since these are handled elsewhere
2. change all data structures to be sync datastructures and not concurrent ones, this introduces a bunch of extra changes  due to mut lifetime references

All relevant tests are also copied, so we can be sure this behaves the same as prior dice in terms of algorithm.

This is not the end state. We can do more simplifications to reduce complexity of the storage, but I'll introduce them slowly once I get new dice up and running

Reviewed By: stepancheg

Differential Revision: D43760287

fbshipit-source-id: cf360da87962d52761b4a11856aa834a555e51b7