Distinguish DiceValue validity by types
Summary:
Create separate DiceValue types of MaybeValid vs regular always valid DiceValue

Rest of it is boilerplate propagation and that automatically results in handling of transients.

There is a slight difference in new vs old Dice, where old DICE is capable to doing early cutoff when the same transient value occurs, but that's a minor incrementality issue that isn't worth the complexity.

Reviewed By: stepancheg

Differential Revision: D43824949

fbshipit-source-id: 92be24ae333bdec9da4d55d145d764852ababd8b