Add ability to synchronously complete a dice task
Summary:
Add the ability to complete a dice task outside the spawn.

This requires a sync function to complete the task that is synchronized against the spawned function that completes the task.

This is mostly done via the state. where we introduce the new state `sync` that serves as spinlock state.
We also add a project state that provides better diagnostics that can occur simultaneously along other states.

Reviewed By: cjhopman

Differential Revision: D43741043

fbshipit-source-id: 3ccd0c8805361c342df41630dacac2378569e3d0