buck2: test: delete a bunch of dead code
Summary:
The only purpose for identifiers is here is for tempdirs, which we don't
actually use in tests because they break many tests. Delete all of this.

A side effect here (and in fact the goal of this stack) is that we no longer
send a random `test-N` to RE in the action key.

Note that this doesn't preclude re-adding a tempdir for tests later, but
it decouples it from the identifier.

Reviewed By: scottcao

Differential Revision: D43905188

fbshipit-source-id: 414d14da8a34f24b3daab299fb8a3804c60c4351