Reland: [buck2][intellij project] Fix module deps not including exported deps for some rule types
Summary:
Redo of D43961798 with the following change: P648553128

I was an idiot and didn't retest after doing a refactor. We need to continue using `resolved_attrs` because `resources` attr may point to a target that needs to be resolved.

And, unfortunately, the failing CI job did not block land because the job doesn't support BRR: https://fb.workplace.com/groups/389567283161794/posts/556288093156378

Reviewed By: jiawei-lyu

Differential Revision: D43987493

fbshipit-source-id: cb2715a0ab4a67b8bf67bb25eb7778ac59267d0e