Remove do_not_build_only_native_code
Summary:
Prior to D42801110, we were matching on the actual configured objects to determine if we should filter something out of an `android_instrumentation_apk` (because it was already present in the APK-under-test).

That diff makes it so that we are actually just comparing the `raw_target`, and so it doesn't matter if we build our Java libs and Android resources in a different cpu configuration (we'll still strip the right libs, and build the same thing).

The advantage of this is that we can use `_build_only_native_code` on our APK-under-test, which reduces the size of analysis by quite a bit.

Reviewed By: jselbo

Differential Revision: D43946445

fbshipit-source-id: 6703efcd7a05d2e3caea1aea4d2f3487df957e57