buck2: log release timestamp in metadata
Summary:
I want to have a dedicated field for "this is a release build from this exact
timestamp", so this is what this adds. The reason I want this is so that I can
omit reporting soft errors & panics from non-release builds and old builds (so
that we can actually close them).

Differential Revision: D44020115

fbshipit-source-id: a7c4d5096ee814b61c9a98755a4af9279a69453c