buck2: forkserver: introduce a "StatusDecoder" that can overwrite the status
Summary:
I want to use this when we have Miniperf. Miniperf will actually return *its
own exit code*, and then return the status of the underlying command via a
separate file (ideally it'd be a pipe, but for reasons explained later in this
stack, we cannot do that).

This will let us hook that in.

Reviewed By: stepancheg

Differential Revision: D44022936

fbshipit-source-id: fe4d1e5a3c301f0c093181b061e1c8b9beb065b5